requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
Deprecated==1.2.14
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
jwcrypto==1.5.1
oauthlib==3.2.2
packaging==23.2
pipreqs==0.4.13
pycparser==2.21
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
yarg==0.1.9
